(refactor) 인수테스트 클래스에서 공통으로 사용하기 위한 베이스 클래스를 분리하라
