+++ src/test/java/com/camping/legacy/test/atdd/AcceptanceTestBase.java
@@ -1,0 +1,22 @@
+package com.camping.legacy.test.atdd;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+/**
+ * 인수 테스트에 필요한 설정을 관리하는 base 클래스
+ */
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public abstract class AcceptanceTestBase {
+
+    @LocalServerPort
+    private int port;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port; // 동적 포트 설정
+        RestAssured.baseURI = "http://localhost"; // 기본 호스트 설정
+    }
+}
